requirements.txt
Flask~=2.3.2
PyMySQL~=1.0.3
setuptools~=67.8.0
python-dotenv~=1.0.0
WTForms~=3.0.1
PyJWT~=2.7.0
SQLAlchemy~=2.0.16